Refactor `geom` library into `geom` and `util`
