+++ LEGO1/lego/sources/util/legomesh.cpp
