+++ LEGO1/lego/sources/util/legosphere.cpp
