+++ LEGO1/lego/sources/util/legovertex.cpp
